+++ src/main/java/com/epam/spring/homework3/payments/controller/dto/AccountDto.java
@@ -10,10 +10,10 @@ import java.time.LocalDate;
 @Builder
 public class AccountDto {
 
-    private int accountId;
-    private int userId;
-    private double creditLimit;
-    private double balance;
+    private Long accountId;
+    private Long userId;
+    private Double creditLimit;
+    private Double balance;
     private String name;
     private String number;
     private String currency;
