+++ src/main/java/com/epam/spring/homework3/payments/controller/dto/PaymentDto.java
@@ -9,13 +9,14 @@ import java.time.LocalDate;
 @Builder
 public class PaymentDto {
 
-    private int paymentId;
-    private int accountIdFrom;
-    private int accountIdTo;
-    private double amount;
+    private Long paymentId;
+    private Long accountIdFrom;
+    private Long accountIdTo;
+    private Double amount;
     private String number;
     private String description;
     private String currency;
     private LocalDate date;
+
 
 }
