+++ src/main/java/com/epam/spring/homework3/payments/model/Account.java
@@ -9,10 +9,10 @@ import java.time.LocalDate;
 @Builder
 public class Account {
 
-    private int accountId;
-    private int userId;
-    private double creditLimit;
-    private double balance;
+    private Long accountId;
+    private Long userId;
+    private Double creditLimit;
+    private Double balance;
     private String name;
     private String number;
     private String currency;
