+++ src/main/java/com/epam/spring/homework3/payments/repository/UserRepository.java
@@ -8,9 +8,9 @@ import java.util.List;
 public interface UserRepository {
 
     User createUser(User user);
-    User getUser(int userId);
-    User updateUser(int userId, User updatedUser);
-    void deleteUser(int userId);
+    User getUser(Long userId);
+    User updateUser(Long userId, User updatedUser);
+    void deleteUser(Long userId);
     List<User> listUsers();
 
 }
