+++ src/main/java/com/epam/spring/homework3/payments/service/PaymentService.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface PaymentService {
 
     PaymentDto createPayment(PaymentDto paymentDto);
-    PaymentDto getPayment(int paymentId);
-    PaymentDto updatePayment(int paymentId, PaymentDto updatedPaymentDto);
-    void deletePayment(int paymentId);
+    PaymentDto getPayment(Long paymentId);
+    PaymentDto updatePayment(Long paymentId, PaymentDto updatedPaymentDto);
+    void deletePayment(Long paymentId);
     List<PaymentDto> listPayments();
 
 }
