+++ src/main/java/com/epam/spring/homework3/payments/service/UserService.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface UserService {
 
     UserDto createUser(UserDto userDto);
-    UserDto getUser(int userId);
-    UserDto updateUser(int userId, UserDto userDto);
-    void deleteUser(int userId);
+    UserDto getUser(Long userId);
+    UserDto updateUser(Long userId, UserDto userDto);
+    void deleteUser(Long userId);
     List<UserDto> listUsers();
 
 }
